Moved data models to a separate file
