+++ Core/Entities/UserSearchResponse.swift
@@ -1,0 +1,12 @@
+//
+//  UserSearchResponse.swift
+//  N11PeopleProject
+//
+//  Created by irem.karakaplan on 21.08.2025.
+//
+
+import Foundation
+
+struct UserSearchResponse: Codable {
+    let items: [User]
+}
